Get a working libbitcoinscript.so build.
This moves a bunch of functions around and, with minimal changes
generates a linkable (though not useful) libbitcoinscript.so.
