+++ src/script/bitcoinscript.h
@@ -1,6 +1,11 @@
 #ifndef H_BITCOIN_BITCOINSCRIPT
 #define H_BITCOIN_BITCOINSCRIPT
 
+#include <vector>
+
+class CScript;
+class CTransaction;
+
 /** Script verification flags */
 enum
 {
@@ -26,5 +31,7 @@ static const unsigned int MANDATORY_SCRIPT_VERIFY_FLAGS = SCRIPT_VERIFY_P2SH;
 static const unsigned int STANDARD_SCRIPT_VERIFY_FLAGS = MANDATORY_SCRIPT_VERIFY_FLAGS |
                                                          SCRIPT_VERIFY_STRICTENC |
                                                          SCRIPT_VERIFY_NULLDUMMY;
+
+bool EvalScript(std::vector<std::vector<unsigned char> >& stack, const CScript& script, const CTransaction& txTo, unsigned int nIn, unsigned int flags, int nHashType);
 
 #endif
